+++ src/solver/thistlewaite/group_two/table.rs
@@ -23,7 +23,7 @@ const SIZE: usize = N_SIZE * M_SIZE;
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Table {
     /// The underlying data, where each index represents a cube state and its depth from the solved state.
-    data: Vec<usize>,
+    data: Vec<u8>,
 }
 
 impl Table {
@@ -33,7 +33,7 @@ impl Table {
     }
 
     /// Returns the number of moves the given cube is, from being in G2.
-    pub fn depth(&self, cube: &Cube) -> usize {
+    pub fn depth(&self, cube: &Cube) -> u8 {
         self.data[idx(cube)]
     }
 }
@@ -53,7 +53,7 @@ fn g2() -> Table {
     // Perform a breadth first search outward from the 96 starting states, applying all the valid G2 moves; the
     // first time a state is reached is guaranteed to be its shortest depth.
     Table {
-        data: bfs_from(Group::Two.moves(), &seeds, DEPTH, SIZE, idx, idx),
+        data: bfs_from(Group::Two.moves(), &seeds, DEPTH as u8, SIZE, idx, idx),
     }
 }
